requirement.txt
streamlit
torch
torchvision
opencv-python
numpy
Pillow
tqdm
yt_dlp
